finsearch/requirements.txt
streamlit>=1.35.0
python-dotenv>=1.0.0
sentence-transformers>=2.7.0
faiss-cpu>=1.8.0
transformers>=4.40.0
torch>=2.2.0
openai>=1.30.0
pymupdf>=1.24.0
pandas>=2.2.0
numpy>=1.26.0
rank-bm25>=0.2.2
